+++ server/controllers/orgController.js
@@ -6,6 +6,7 @@ const sequelize = require('../db/sequelize').sequelize;
 const token = require('../token/token');
 
 
+
 // right now, this function doesn't seem to fetch all subordinates.  Example - George Nacouzi doesn't show up under Ethan
 function show(req, res) {
   const emailAddress = req.params.emailAddress;
@@ -100,11 +101,66 @@ function getTeamFteList(req, res) {
     });
 }
 
+function getManagementOrgStructure(req, res) {
+
+  const decodedToken = token.decode(req.header('X-Token'), res);
+
+  console.log('decoded token for get management org structure:');
+  console.log(decodedToken);
+
+  const emailAddress = req.params.emailAddress;
+  
+  sequelize.query(`EXECUTE resources.ManagementOrgStructure :emailAddress`, {replacements: {emailAddress: emailAddress}, type: sequelize.QueryTypes.SELECT})
+  .then(org => {
+    res.json(org);
+  })
+  .catch(error => {
+    res.status(400).json({
+      title: 'Error (in catch)',
+      error: {message: error}
+    })
+  });
+}
+
+function getOrgStructureDrillDown(req, res) {
+  // Retrieves muptiple levels of employees and managers
+  const emailAddress = req.params.emailAddress;
+  sequelize.query('EXECUTE resources.OrgStructureDrillDown :emailAddress', {replacements: {emailAddress: emailAddress}, type: sequelize.QueryTypes.SELECT})
+    .then(org => {
+      res.json(org);
+    })
+    .catch(error => {
+      res.status(400).json({
+        title: 'Error (in catch)',
+        error: {message: error}
+      })
+    });
+}
+
+function getOrgStructureDrillUp(req, res) {
+  // Retrieves management chain above the employee
+  const emailAddress = req.params.emailAddress;
+  sequelize.query('EXECUTE resources.OrgStructureDrillUp :emailAddress', {replacements: {emailAddress: emailAddress}, type: sequelize.QueryTypes.SELECT})
+    .then(org => {
+      res.json(org);
+    })
+    .catch(error => {
+      res.status(400).json({
+        title: 'Error (in catch)',
+        error: {message: error}
+      })
+    });
+}
+
+
 module.exports = {
   show: show,
   getSubordinatesFlat: getSubordinatesFlat,
   getTeamList: getTeamList,
   getEmployeeList: getEmployeeList,
   getOrgFtes: getOrgFtes,
-  getTeamFteList: getTeamFteList
+  getTeamFteList: getTeamFteList,
+  getManagementOrgStructure: getManagementOrgStructure,
+  getOrgStructureDrillDown: getOrgStructureDrillDown,
+  getOrgStructureDrillUp: getOrgStructureDrillUp
 }