+++ src/app/_shared/services/api-data/api-data-org.service.ts
@@ -47,4 +47,24 @@ export class ApiDataOrgService {
       .map((response: Response) => response.json());
   }
 
+  getManagementOrgStructure(emailAddress: string): Observable<any> {
+    const headers = new Headers({'Content-Type': 'application/json', 'X-Token': this.cacheService.token.signedToken});
+    const options = new RequestOptions({headers: headers});
+    return this.http.get(`/api/org/getManagementOrgStructure/${emailAddress}`, options)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
+  getOrgStructureDrillDown(emailAddress: string): Observable<any> {
+    return this.http.get(`/api/org/getOrgStructureDrillDown/${emailAddress}`)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
+  getOrgStructureDrillUp(emailAddress: string): Observable<any> {
+    return this.http.get(`/api/org/getOrgStructureDrillUp/${emailAddress}`)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
 }